requirements-test.txt
# Testing dependencies
pytest==8.3.3
pytest-asyncio==0.25.0
pytest-mock==3.14.0
pytest-cov==6.0.0
pytest-tornado==0.8.1
coverage==7.6.9
mock==5.1.0